ci: make reset more robust

+++ reset.sh
@@ -1,2 +1,8 @@
-git tag -d v0.1.0
-git reset --hard HEAD~1
+set -e
+
+N_TAG=$(git tag | wc -l)
+git tag | xargs -r git tag -d
+
+if [[ $N_TAG -ne 0 ]]; then
+    git reset --hard HEAD~1
+fi
